requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
snowflake-snowpark-python>=1.9.0
snowflake-connector-python>=3.0.0
